response-playbooks/requirements.txt
elasticsearch>=8.0.0
flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
